fix(files): fix personal file uploads
