--- libraries/Iridium/types.ts
@@ -1,6 +1,0 @@
-export type User = {
-  id: string
-  name: string
-  avatar: string
-  metadata: { [key: string]: any }
-}
